Renamed pytrackmain to pytrack worker
